Added useful debugging statements
Write-Verbose and Write-Debug so will only be visible when required and
specifically asked for.

The Write-Debug immediately before the root directory is
removed is very handy to allow a developer to inspect what was done before all
the evidence is detroyed.

Since the parameters passed to 7zip come from a couple of places, outputting the
final set with Write-Verbose is useful for debugging.